+++ discord/claim.go
@@ -4,7 +4,9 @@ import (
 	"bot/database"
 	"bot/query"
 	"fmt"
+	"math/rand"
 	"strings"
+	"time"
 
 	"github.com/andersfylling/disgord"
 )
@@ -15,6 +17,31 @@ var char query.CharStruct
 // enableClaim is used to make the new character claimable by the user
 func enableClaim(in query.CharStruct) {
 	char = in
+}
+
+func drop(data *disgord.MessageCreate) {
+	query := query.CharSearchByPopularity(rand.New(rand.NewSource(time.Now().UnixNano())).Intn(conf.MaxChar))
+	enableClaim(query)
+	printDrop(data, query.Page.Characters[0].Image.Large)
+}
+
+func printDrop(data *disgord.MessageCreate, image string) {
+	// Sends the message
+	client.CreateMessage(
+		ctx,
+		data.Message.ChannelID,
+		&disgord.CreateMessageParams{
+			Embed: &disgord.Embed{
+				Title:       "A new character appeared",
+				Description: fmt.Sprintf("use %sclaim to get this character for yourself", conf.Prefix),
+				Image: &disgord.EmbedImage{
+					URL: image,
+				},
+				Timestamp: data.Message.Timestamp,
+				Color:     0xF2FF2E,
+			},
+		},
+	)
 }
 
 // claim is used to claim a waifu and add it to your database
